requirements.txt
# Web Application Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Explainability
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
numpy>=1.24.0

# Utilities
tqdm>=4.65.0
pandas>=2.0.0
scikit-learn>=1.3.0
